requirements.txt
tweepy
pushbullet.py
paho-mqtt
#curses-menu
scrollphat
